+++ src/vm/x64.rs
@@ -45,21 +45,21 @@ pub fn init_x64(
     // map direct map region
     for i in 0..DIRECT_MAP_PML4_ENTRIES_COUNT {
         let entry_val =
-            (DIRECT_MAP_PDPT.as_u64() + i as u64 * PAGE_TABLE_SIZE) | PTE_PRESENT | PTE_RW;
+            (DIRECT_MAP_PDPT.as_u64() + i as u64 * PAGE_TABLE_SIZE as u64) | PTE_PRESENT | PTE_RW;
         let entry_addr =
             GuestAddress(DIRECT_MAP_PML4.as_u64() + ((DIRECT_MAP_PML4_OFFSET + i) * 8) as u64);
         boot_mem.write_slice(&entry_val.to_le_bytes(), entry_addr)?;
     }
 
     for i in 0..DIRECT_MAP_PDPT_COUNT * PAGE_TABLE_ENTRIES {
-        let pd_phys = DIRECT_MAP_PD.as_u64() + i as u64 * PAGE_TABLE_SIZE;
+        let pd_phys = DIRECT_MAP_PD.as_u64() + i as u64 * PAGE_TABLE_SIZE as u64;
         let entry_val = pd_phys | PTE_PRESENT | PTE_RW;
         let entry_addr = GuestAddress(DIRECT_MAP_PDPT.as_u64() + (i * 8) as u64);
         boot_mem.write_slice(&entry_val.to_le_bytes(), entry_addr)?;
     }
 
     for i in 0..DIRECT_MAP_PD_COUNT * PAGE_TABLE_ENTRIES {
-        let phys = i as u64 * PAGE_SIZE;
+        let phys = i as u64 * PAGE_SIZE as u64;
         let entry_val = phys | PTE_PRESENT | PTE_RW | PTE_PS;
         let entry_addr = GuestAddress(DIRECT_MAP_PD.as_u64() + (i * 8) as u64);
         boot_mem.write_slice(&entry_val.to_le_bytes(), entry_addr)?;
@@ -68,19 +68,19 @@ pub fn init_x64(
     // map kernel code region
     let kernel_pml4_val = KERNEL_CODE_PDPD.as_u64() | PTE_PRESENT | PTE_RW;
     let kernel_pml4_addr =
-        GuestAddress(DIRECT_MAP_PML4.as_u64() + KERNEL_CODE_VIRT.pml4_index() * 8);
+        GuestAddress(DIRECT_MAP_PML4.as_u64() + (KERNEL_CODE_VIRT.pml4_index() * 8) as u64);
     boot_mem.write_slice(&kernel_pml4_val.to_le_bytes(), kernel_pml4_addr)?;
 
     for i in 0..2 {
-        let pd_phys = KERNEL_CODE_PD.as_u64() + (i as u64 * PAGE_TABLE_SIZE);
+        let pd_phys = KERNEL_CODE_PD.as_u64() + (i as u64 * PAGE_TABLE_SIZE as u64);
         let entry_val = pd_phys | PTE_PRESENT | PTE_RW;
         let entry_addr =
-            GuestAddress(KERNEL_CODE_PDPD.as_u64() + (KERNEL_CODE_VIRT.pdpt_index() + i) * 8);
+            GuestAddress(KERNEL_CODE_PDPD.as_u64() + ((KERNEL_CODE_VIRT.pdpt_index() + i) * 8) as u64);
         boot_mem.write_slice(&entry_val.to_le_bytes(), entry_addr)?;
     }
 
     for i in 0..PAGE_TABLE_ENTRIES {
-        let phys = KERNEL_CODE_PHYS.add(i as u64 * PAGE_SIZE).as_u64();
+        let phys = KERNEL_CODE_PHYS.add(i * PAGE_SIZE).as_u64();
         let entry_val = phys | PTE_PRESENT | PTE_RW | PTE_PS;
         let entry_addr = GuestAddress(KERNEL_CODE_PD.as_u64() + (i * 8) as u64);
         boot_mem.write_slice(&entry_val.to_le_bytes(), entry_addr)?;
